# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Todo Camisetas API",
        "description": "API B2B para gestión de camisetas y clientes",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Servidor local Docker"
        }
    ],
    "paths": {
        "/api/camisetas": {
            "get": {
                "tags": [
                    "Camisetas"
                ],
                "summary": "Listar todas las camisetas",
                "operationId": "d5e9027b7ea9dd37ebe7ce5b97881177",
                "responses": {
                    "200": {
                        "description": "Lista de camisetas con sus tallas"
                    }
                }
            },
            "post": {
                "tags": [
                    "Camisetas"
                ],
                "summary": "Crear una camiseta",
                "operationId": "c9a38b0a0a0ec748f6c48887f08206ee",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "titulo",
                                    "club",
                                    "pais",
                                    "tipo",
                                    "color",
                                    "precio",
                                    "codigo_producto"
                                ],
                                "properties": {
                                    "titulo": {
                                        "type": "string",
                                        "example": "Camiseta Local 2025"
                                    },
                                    "club": {
                                        "type": "string",
                                        "example": "Selección Chilena"
                                    },
                                    "pais": {
                                        "type": "string",
                                        "example": "Chile"
                                    },
                                    "tipo": {
                                        "type": "string",
                                        "example": "Local"
                                    },
                                    "color": {
                                        "type": "string",
                                        "example": "Rojo y Azul"
                                    },
                                    "precio": {
                                        "type": "number",
                                        "example": 45000
                                    },
                                    "precio_oferta": {
                                        "type": "number",
                                        "example": 38000,
                                        "nullable": true
                                    },
                                    "detalles": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "codigo_producto": {
                                        "type": "string",
                                        "example": "SCL2025L"
                                    },
                                    "tallas": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Camiseta creada"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/camisetas/{id}": {
            "get": {
                "tags": [
                    "Camisetas"
                ],
                "summary": "Ver una camiseta por ID",
                "operationId": "1adc1182bb5738b156a92364eabc2844",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Camiseta encontrada"
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Camisetas"
                ],
                "summary": "Actualizar una camiseta",
                "operationId": "423e3360b1e6917c7ddd97815131398f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "titulo": {
                                        "type": "string"
                                    },
                                    "precio": {
                                        "type": "number"
                                    },
                                    "precio_oferta": {
                                        "type": "number",
                                        "nullable": true
                                    },
                                    "tallas": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Camiseta actualizada"
                    },
                    "404": {
                        "description": "No encontrada"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Camisetas"
                ],
                "summary": "Eliminar una camiseta",
                "operationId": "f798d8ac4eaa052e0f3d9e90cace7718",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Camiseta eliminada"
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                }
            }
        },
        "/api/camisetas/{id}/precio": {
            "get": {
                "tags": [
                    "Camisetas"
                ],
                "summary": "Precio final según cliente",
                "operationId": "65a4a88ff464e3e3e9c7aff3c06a7fe3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "cliente_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Precio final calculado"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            }
        },
        "/api/clientes/{cliente_id}/camisetas": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Camisetas con precio final por cliente",
                "operationId": "5ffdbfe3efa2bf5499fea7aa9c1b0224",
                "parameters": [
                    {
                        "name": "cliente_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista con precio_final calculado"
                    },
                    "404": {
                        "description": "Cliente no encontrado"
                    }
                }
            }
        },
...
    },
...
}